Fix: created helper test function to clean up tests and to be used for others if needed
