+++ lib/supabase/middleware/middleware.test.ts
@@ -3,13 +3,7 @@
  */
 
 import { NextRequest } from 'next/server';
-import { updateSession } from './middleware';
-import { createClient } from '@supabase/supabase-js';
-
-const supabase = createClient(
-  process.env.NEXT_PUBLIC_SUPABASE_URL!,
-  process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!,
-);
+import { createAuthedRequest, updateSession } from './middleware';
 
 describe('Middleware', () => {
   it('redirects unauthenticated user back to /', async () => {
@@ -19,22 +13,11 @@ describe('Middleware', () => {
     expect(response.headers.get('location')).toBe('https://localhost:4000/');
   });
   it('redirects user to go to /onboarding if their onboarding is set as false', async () => {
-    const {
-      data: { session },
-    } = await supabase.auth.signInWithPassword({
+    const response = await createAuthedRequest({
       email: 'onboardingfalse@test.com',
       password: 'password',
+      path: '/dashboard',
     });
-
-    const cookieHeader = `sb-access-token=${session?.access_token}; sb-refresh-token=${session?.refresh_token}`;
-
-    const request = new NextRequest('https://localhost:4000/dashboard', {
-      headers: new Headers({
-        cookie: cookieHeader,
-      }),
-    });
-
-    const response = await updateSession(request);
     expect(response.status).toBe(307);
     expect(response.headers.get('location')).toBe(
       'https://localhost:4000/onboarding',
@@ -42,42 +25,20 @@ describe('Middleware', () => {
   });
 
   it('allows user to go to /dashboard if their onboarding is set as true', async () => {
-    const {
-      data: { session },
-    } = await supabase.auth.signInWithPassword({
+    const response = await createAuthedRequest({
       email: 'onboardingtrue@test.com',
       password: 'password',
+      path: '/dashboard',
     });
-
-    const cookieHeader = `sb-access-token=${session?.access_token}; sb-refresh-token=${session?.refresh_token}`;
-
-    const request = new NextRequest('https://localhost:4000/dashboard', {
-      headers: new Headers({
-        cookie: cookieHeader,
-      }),
-    });
-
-    const response = await updateSession(request);
     expect(response.status).toBe(200);
   });
 
   it('does not redirect user away from /onboarding if they want to go back and edit their profile', async () => {
-    const {
-      data: { session },
-    } = await supabase.auth.signInWithPassword({
+    const response = await createAuthedRequest({
       email: 'onboardingtrue@test.com',
       password: 'password',
+      path: '/onboarding',
     });
-
-    const cookieHeader = `sb-access-token=${session?.access_token}; sb-refresh-token=${session?.refresh_token}`;
-
-    const request = new NextRequest('https://localhost:4000/onboarding', {
-      headers: new Headers({
-        cookie: cookieHeader,
-      }),
-    });
-
-    const response = await updateSession(request);
     expect(response.status).toBe(200);
   });
